requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pyhd8ed1ab_0
annotated-types=0.6.0=pyhd8ed1ab_0
anyio=4.2.0=pyhd8ed1ab_0
archspec=0.2.2=pyhd8ed1ab_0
boltons=23.1.1=pyhd8ed1ab_0
brotli-python=1.1.0=py39h99910a6_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2023.11.17=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=py39ha55989b_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
conda=23.11.0=py39hcbf5309_1
conda-libmamba-solver=23.12.0=pyhd8ed1ab_0
conda-package-handling=2.2.0=pyh38be061_0
conda-package-streaming=0.9.0=pyhd8ed1ab_0
distro=1.9.0=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
fmt=10.2.1=h181d51b_0
google-ai-generativelanguage=0.4.0=pypi_0
google-api-core=2.15.0=pypi_0
google-auth=2.27.0=pypi_0
google-generativeai=0.3.2=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpcio=1.60.0=pypi_0
grpcio-status=1.60.0=pypi_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.2=pyhd8ed1ab_0
httpx=0.26.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.6=pyhd8ed1ab_0
intel-openmp=2024.0.0=h57928b3_49841
joblib=1.3.2=pyhd8ed1ab_0
jsonpatch=1.33=pyhd8ed1ab_0
jsonpointer=2.4=py39hcbf5309_3
krb5=1.21.2=heb0366b_0
libarchive=3.7.2=h313118b_1
libblas=3.9.0=21_win64_mkl
libcblas=3.9.0=21_win64_mkl
libcurl=8.5.0=hd5e4a3a_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=hcfcfb64_2
liblapack=3.9.0=21_win64_mkl
libmamba=1.5.6=h3f09ed1_0
libmambapy=1.5.6=py39h2690a07_0
libsolv=0.7.27=h12be248_0
libsqlite=3.44.2=hcfcfb64_0
libssh2=1.11.0=h7dfc565_0
libxml2=2.12.4=hc3477c8_1
libzlib=1.2.13=hcfcfb64_5
lz4-c=1.9.4=hcfcfb64_0
lzo=2.10=he774522_1000
mamba=1.5.6=py39hca8391b_0
menuinst=2.0.2=py39h99910a6_0
mkl=2024.0.0=h66d3029_49657
nltk=3.8.1=pyhd8ed1ab_0
numpy=1.26.3=py39hddb5d58_0
openai=1.10.0=pyhd8ed1ab_0
openssl=3.2.0=hcfcfb64_1
packaging=23.2=pyhd8ed1ab_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pyhd8ed1ab_0
pluggy=1.4.0=pyhd8ed1ab_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.2=pypi_0
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybind11-abi=4=hd8ed1ab_3
pycosat=0.6.6=py39ha55989b_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=2.5.3=pyhd8ed1ab_0
pydantic-core=2.14.6=py39hf21820d_1
pysocks=1.7.1=pyh0701188_6
python=3.9.18=h4de0772_1_cpython
python_abi=3.9=4_cp39
ratelimit=2.2.1=pyhd8ed1ab_0
regex=2023.12.25=py39ha55989b_0
reproc=14.2.4.post0=hcfcfb64_1
reproc-cpp=14.2.4.post0=h63175ca_1
requests=2.31.0=pyhd8ed1ab_0
rouge-score=0.1.2=pyhd8ed1ab_0
rsa=4.9=pypi_0
ruamel.yaml=0.18.5=py39ha55989b_0
ruamel.yaml.clib=0.2.7=py39ha55989b_2
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
tbb=2021.11.0=h91493d7_1
tiktoken=0.5.2=py39hca44cb7_0
tk=8.6.13=h5226925_1
tqdm=4.66.1=pyhd8ed1ab_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=pyhd8ed1ab_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wheel=0.42.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xz=5.2.6=h8d14728_0
yaml-cpp=0.8.0=h63175ca_0
zstandard=0.22.0=py39h95af829_0
zstd=1.5.5=h12be248_0
